requirements.txt
discord.py
Flask
PuNaCl